fix: move tests into DatabendIT
